backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-multipart==0.0.20
bcrypt==4.0.1
pydantic-settings==2.10.1
cryptography==45.0.6
